+++ CHANGELOG.md
@@ -11,3 +11,4 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Figure axes are not resized by default any more when plotting a polygon
+- Use ``pylab.show()`` rather than IPython in examples